Small change to calculator_refactor.rb

+++ RB101/lesson_02/calculator_refactor.rb
@@ -84,6 +84,7 @@ loop do # main loop
   operator = ''
   loop do
     operator = Kernel.gets().chomp()
+    
     if valid_operator?(operator)
       break
     else
@@ -101,7 +102,7 @@ loop do # main loop
            when '3'
              number1.to_i() * number2.to_i()
            when '4'
-             number1.to_f() / number2.to_f()
+             number1.to_i() / number2.to_i()
            end
 
   prompt("The result is #{result}")
